Simplify main handler, add handler factory
The handler factory helps other handlers to gain access to other
handlers if they cannot handle a case well. Usually that would be either
serving an HTTP500 server error or redirecting with the help of the CAS
handler.

Because the previous structure was handlers-embed-other-handlers-which-
also-embed-handlers, this commit breaks this structure with help of the
factory. Otherwise, a handler might be called twice which would not make
sense and also leads to errors.